% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContrasts.R
\name{reverseComparison}
\alias{reverseComparison}
\title{Reverse Comparison}
\usage{
reverseComparison(co)
}
\arguments{
\item{co}{Either a named character vector or a named list as input.}
}
\value{
A named list of character vector containing the reversed comparisons
}
\description{
Reverses a comparison or a list of comparisons.
}
